Changes in style and isolation of code on new methods
