+++ COMETwebapp/Primitives/Cube.cs
@@ -26,7 +26,7 @@ namespace COMETwebapp.Primitives
     /// <summary>
     /// Cube primitive type
     /// </summary>
-    public class Cube : PositionablePrimitives
+    public class Cube : PositionablePrimitive
     {
         /// <summary>
         /// The width of the cube
