+++ COMETwebapp/Primitives/Cylinder.cs
@@ -26,7 +26,7 @@ namespace COMETwebapp.Primitives
     /// <summary>
     /// Cylinder primitive type
     /// </summary>
-    public class Cylinder : PositionablePrimitives
+    public class Cylinder : PositionablePrimitive
     {
         /// <summary>
         /// Radius of the <see cref="Cylinder"/>
