+++ COMETwebapp/Primitives/IShapeFactory.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="Sphere.cs" company="RHEA System S.A.">
+// <copyright file="IShapeFactory.cs" company="RHEA System S.A.">
 //    Copyright (c) 2022 RHEA System S.A.
 //
 //    Author: Sam Gerené, Alex Vorobiev, Alexander van Delft, Jaime Bernar
@@ -21,45 +21,13 @@
 //    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
+
 namespace COMETwebapp.Primitives
 {
-    /// <summary>
-    /// Sphere primitive type
-    /// </summary>
-    public class Sphere : PositionablePrimitive
+    using CDP4Common.EngineeringModelData;
+
+    public interface IShapeFactory
     {
-        /// <summary>
-        /// Basic type name
-        /// </summary>
-        public override string Type { get; protected set; } = "Sphere";
-
-        /// <summary>
-        /// The radius of the <see cref="Sphere"/>
-        /// </summary>
-        public double Radius { get; }
-
-        /// <summary>
-        /// Initializes a new instance of <see cref="Sphere"/> class
-        /// </summary>
-        /// <param name="radius"></param>
-        public Sphere(double radius)
-        {
-            this.Radius = radius;
-        }
-
-        /// <summary>
-        /// Initializes a new instance of <see cref="Sphere"/> class
-        /// </summary>
-        /// <param name="x">position along the x axis</param>
-        /// <param name="y">position along the y axis</param>
-        /// <param name="z">position along the z axis</param>
-        /// <param name="radius">the radius of the sphere</param>
-        public Sphere(double x, double y, double z, double radius)
-        {
-            this.X = x;
-            this.Y = y;
-            this.Z = z;
-            this.Radius = radius;
-        }
+        bool TryGetPrimitiveFromElementUsageParameter(ElementUsage elementUsage, out Primitive basicShape);
     }
 }
